Mac build fix
Review URL: http://codereview.chromium.org/18122


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@8138 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/waitable_event.h
@@ -136,8 +136,8 @@ class WaitableEvent {
   // second element is the index of the WaitableEvent in the original,
   // unsorted, array.
   typedef std::pair<WaitableEvent*, size_t> WaiterAndIndex;
-  static unsigned EnqueueMany(WaiterAndIndex* waitables,
-                              size_t count, Waiter* waiter);
+  static size_t EnqueueMany(WaiterAndIndex* waitables,
+                            size_t count, Waiter* waiter);
 
   Lock lock_;
   bool signaled_;
